requirements.txt
flask
pydantic-ai
python-dotenv
streamlit